活动比赛绘制了一半，暂时存放
