+++ nonebot_plugin_splatoon3/static_data_getter.py
@@ -3,7 +3,7 @@ import httpx
 from bs4 import BeautifulSoup
 from nonebot.log import logger
 from .image_db import imageDB
-from .image_processer import get_cf_file_url, get_file_url
+from .image_processer_tools import get_file_url
 from ._class import WeaponData, ImageInfo
 from .translation import (
     dict_weapon_sub_trans,
