+++ test.py
@@ -51,10 +51,13 @@ from nonebot_plugin_splatoon3.static_data_getter import reload_weapon_info
 
 
 # 测试新版随机武器
-plain_text = "随机武器 nice弹"
-res = get_random_weapon_image(plain_text)
-res.show()
+# plain_text = "随机武器 nice弹"
+# res = get_random_weapon_image(plain_text)
+# res.show()
 
+# 测试活动
+res = get_events_image()
+res.show()
 
 # 写出武器翻译字典
 # write_weapon_trans_dict()
